- fix: removed circular dependencies
